Module google_cloudlatencytest2::api

Structs

AggregatedStats
There is no detailed description.

AggregatedStatsReply
There is no detailed description.

Cloudlatencytest
Central instance to access all Cloudlatencytest related resource activities

DoubleValue
There is no detailed description.

IntValue
There is no detailed description.

Stats
There is no detailed description.

StatsReply
There is no detailed description.

StatscollectionMethods
A builder providing access to all methods supported on statscollection resources. It is not used directly, but through the Cloudlatencytest hub.

StatscollectionUpdateaggregatedstatCall
RPC to update the new TCP stats.

StatscollectionUpdatestatCall
RPC to update the new TCP stats.

StringValue
There is no detailed description.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.